backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
PyPDF2==3.0.1
geopy==2.4.1
requests==2.31.0
python-dateutil==2.8.2
reportlab==4.0.7
Pillow==10.1.0
werkzeug==3.0.1
